requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.3
django-cache-url==3.4.0
django-ckeditor==6.3.0
django-js-asset==2.0.0
environs==9.3.5
gunicorn==20.1.0
marshmallow==3.15.0
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.8
python-dotenv==0.20.0
sqlparse==0.4.2
whitenoise==5.3.0
